requirements.txt
opencv-python==4.11.0.86
matplotlib==3.10.3
pillow==11.2.1
torch==2.7.1
torchvision==0.22.1
ultralytics==8.3.156
